refactor: rename internal function file names to utils
