+++ src/functions/addOperator/utils.ts
@@ -2,7 +2,7 @@ import { CalculationString } from '@src/types';
 import { isEndsWith, removeCharactersFromEnd } from '@src/utils';
 
 const toggleMinusOperator = (
-  calculationString: CalculationString
+  calculationString: CalculationString,
 ): CalculationString => {
   if (isEndsWith(calculationString, '-')) {
     return removeCharactersFromEnd(calculationString);
