+++ src/functions/addThousandsSeparator/utils.ts
