+++ src/functions/getResult/utils.ts
