+++ src/functions/togglePoint/utils.ts
